refactor: 카테고리탭 디렉토리 분리
