+++ fe/src/Pages/Home.tsx
@@ -2,7 +2,7 @@ import { useEffect, useState } from "react";
 import styles from "./Home.module.css";
 import { AnimationClass } from "../types/constants";
 import { API_URL } from "../constants";
-import CategoryTab from "../components/CategoryTab";
+import CategoryTab from "../components/CategoryTab/CategoryTab";
 import Main from "../components/Main";
 
 type HomeProps = {
